# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Node.js IDE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FileEntry {
  /** Relative file path (e.g. "src/index.js") */
  path: string;
  /** File name */
  name: string;
  /** File size in bytes */
  size: number;
  updatedAt: string;
}

export interface FileContent {
  path: string;
  content: string;
}

export interface FileInput {
  /** Relative path for the new file (e.g. "hello.js") */
  path: string;
  /** Initial content */
  content?: string;
}

export interface FileUpdate {
  content: string;
}

export interface RenameInput {
  /** New relative path */
  newPath: string;
}

export interface DeleteResult {
  path: string;
  deleted: boolean;
}

export interface ExecuteInput {
  /** Path of file to run */
  filePath: string;
  /** Optional stdin input */
  stdin?: string;
}

export interface ExecuteResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  durationMs: number;
}