# lib/api-zod/src/generated/types/executeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Node.js IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface ExecuteInput {
  /** Path of file to run */
  filePath: string;
  /** Optional stdin input */
  stdin?: string;
}