# lib/api-zod/src/generated/types/fileEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Node.js IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface FileEntry {
  /** Relative file path (e.g. "src/index.js") */
  path: string;
  /** File name */
  name: string;
  /** File size in bytes */
  size: number;
  updatedAt: Date;
}