# lib/api-zod/src/generated/types/fileInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Node.js IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface FileInput {
  /** Relative path for the new file (e.g. "hello.js") */
  path: string;
  /** Initial content */
  content?: string;
}